Move Serializer Code into utilities, add READMEs
