+++ rollbar-payload/src/main/java/com/rollbar/payload/Payload.java
@@ -5,6 +5,7 @@ import com.rollbar.payload.data.Level;
 import com.rollbar.payload.data.Notifier;
 import com.rollbar.payload.data.body.Body;
 import com.rollbar.utilities.ArgumentNullException;
+import com.rollbar.utilities.RollbarSerializer;
 import com.rollbar.utilities.Validate;
 import com.rollbar.utilities.JsonSerializable;
