--- rollbar-payload/src/main/java/com/rollbar/payload/PayloadSerializer.java
@@ -1,15 +1,0 @@
-package com.rollbar.payload;
-
-import com.rollbar.payload.Payload;
-
-/**
- * The interface that encapsulates turning a {@link Payload} into a JSON payload string.
- */
-public interface PayloadSerializer {
-    /**
-     * Turn the {@link Payload} into a json string
-     * @param payload the {@link Payload} to serialize
-     * @return the json payload.
-     */
-    String serialize(Payload payload);
-}
