+++ rollbar-testing/README.md
@@ -1,0 +1,5 @@
+## Overview
+
+`rollbar-testing` contains shared test-code. If you find yourself
+needing the same test helpers across modules you should include them
+here.
