+++ rollbar-utilities/README.md
@@ -1,0 +1,5 @@
+## Overview
+
+`rollbar-utilities` contains code that's shared by the other rollbar
+modules. These are things like Exceptions, serialization, validation,
+and string manipulation code.
